+++ Assets/i5 Toolkit for Unity/Runtime/ServiceCore/Scripts/ServiceManager.cs
@@ -9,7 +9,7 @@ namespace i5.Toolkit.Core.ServiceCore
     /// Manager which administers registered services
     /// These services need to implement the IService interface and do not need to inherit from MonoBehaviour
     /// </summary>
-    public class ServiceManager : MonoBehaviour
+    public class ServiceManager : BaseServiceManager
     {
         private Dictionary<Type, IService> registeredServices = new Dictionary<Type, IService>();
 
@@ -66,30 +66,40 @@ namespace i5.Toolkit.Core.ServiceCore
         public static void RegisterService<T>(T service) where T : IService
         {
             EnsureInstance();
-            if (instance.registeredServices.ContainsKey(typeof(T)))
+            instance.InstRegisterService(service);
+        }
+
+        public override void InstRegisterService<T>(T service)
+        {
+            if (registeredServices.ContainsKey(typeof(T)))
             {
-                i5Debug.LogError("An instance of this service is already registered", instance);
+                i5Debug.LogError("An instance of this service is already registered", this);
                 return;
             }
-            instance.registeredServices.Add(typeof(T), service);
+            registeredServices.Add(typeof(T), service);
 
             if (service is IUpdateableService)
             {
-                instance.updateableServices.Add((IUpdateableService)service);
+                updateableServices.Add((IUpdateableService)service);
             }
 
-            service.Initialize(instance);
+            service.Initialize(this);
         }
 
         public static void RemoveService<T>() where T : IService
         {
             EnsureInstance();
-            if (!instance.inDestroyMode)
+            instance.InstRemoveService<T>();
+        }
+
+        public override void InstRemoveService<T>()
+        {
+            if (!inDestroyMode)
             {
-                if (instance.registeredServices.ContainsKey(typeof(T)))
+                if (registeredServices.ContainsKey(typeof(T)))
                 {
-                    instance.registeredServices[typeof(T)].Cleanup();
-                    instance.registeredServices.Remove(typeof(T));
+                    registeredServices[typeof(T)].Cleanup();
+                    registeredServices.Remove(typeof(T));
                 }
                 else
                 {
@@ -101,17 +111,27 @@ namespace i5.Toolkit.Core.ServiceCore
         public static T GetService<T>() where T : IService
         {
             EnsureInstance();
-            if (!instance.registeredServices.ContainsKey(typeof(T)))
+            return instance.InstGetService<T>();
+        }
+
+        public override T InstGetService<T>()
+        {
+            if (!registeredServices.ContainsKey(typeof(T)))
             {
                 throw new InvalidOperationException("Tried to get unregistered service");
             }
-            return (T)instance.registeredServices[typeof(T)];
+            return (T)registeredServices[typeof(T)];
         }
 
         public static bool ServiceExists<T>() where T : IService
         {
             EnsureInstance();
-            return instance.registeredServices.ContainsKey(typeof(T));
+            return instance.InstServiceExists<T>();
+        }
+
+        public override bool InstServiceExists<T>()
+        {
+            return registeredServices.ContainsKey(typeof(T));
         }
 
         private void Update()